spray can room assets  ( in code ) other minor updates

+++ src/controls.cpp
@@ -43,7 +43,7 @@ void Controls::setMouseButton(int button, int action, int mods)
     }
 }
 
-void Controls::setVRController(int index, const glm::vec3& position, const glm::quat& orientation, uint32_t pressedFlags)
+void Controls::setVRController(int index, const glm::vec3& position, const glm::quat& orientation, uint64_t pressedFlags)
 {
     _vrControllers[index].position = position;
     _vrControllers[index].orientation = orientation;
